backend/requirements.env.txt
# ====================================
# Librerías necesarias para la ejecución del backend
# ====================================
# ------------------------------------
# Librerías servidor
# ------------------------------------
uvicorn == 0.18.3
gunicorn == 20.1.0
fastapi == 0.85.0
websockets == 10.4
passlib[bcrypt] == 1.7.4
pyjwt == 2.5.0
pytest == 7.2.0
# ------------------------------------
# Librerías modelo IA LSC
# ------------------------------------
mediapipe == 0.8.11
opencv-python == 4.6.0.66
tensorflow == 2.10.0
scipy == 1.9.2
seaborn == 0.12.0
scikit-learn == 1.1.2
# ------------------------------------
# Librerías base de datos
# ------------------------------------
tinydb == 4.7.0
tinydb_serialization == 2.1.0
# ------------------------------------
# Librerías de utilidades extra
# ------------------------------------
python-dotenv == 0.21.0
black == 22.10.0
ipykernel == 6.16.0
ipympl == 0.9.2
ipython-autotime == 0.3.1
#protobuf==3.20.*